requirements.txt
# Core Dependencies
python-telegram-bot>=20.0
telethon==1.34.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase>=2.0.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pgvector==0.2.4
psycopg2-binary>=2.9.0

# AI/ML
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.3
langchain-community==0.0.10
tiktoken>=0.5.0
pytesseract==0.3.10
Pillow==10.2.0
open-clip-torch==2.24.0

# AWS & Storage
boto3==1.34.0
botocore>=1.29.0

# Redis & Caching
redis>=5.0.0
hiredis>=2.0.0
upstash-redis==1.1.0

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0

# Utilities
httpx==0.27.0
aiofiles>=23.0.0
cryptography==41.0.7
aiosqlite==0.19.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11